+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [3.2.3] 2019-07-22
+
+### Changed
+
+- Performance improvements for `TemplateCache` and `ConceptCache`
+- Support configurable port for Redis
+- Publish `delete_link` command when a link is deleted
+
 ## [3.2.1] 2019-07-18
 
 ### Changed